ENH: move 'conversions' and 'spreadsheet_custom_functions' to 'server' code
